Trying to get Activity to receive broadcasts.
