+++ .idea/modules.xml
@@ -3,11 +3,9 @@
   <component name="ProjectModuleManager">
     <modules>
       <module fileurl="file://$PROJECT_DIR$/Dispatch.iml" filepath="$PROJECT_DIR$/Dispatch.iml" />
-      <module fileurl="file://$PROJECT_DIR$/Dispatch.iml" filepath="$PROJECT_DIR$/Dispatch.iml" />
-      <module fileurl="file://$PROJECT_DIR$/../../../../NikoPC/Documents/GitHub/Dispatch/Dispatch.iml" filepath="$PROJECT_DIR$/../../../../NikoPC/Documents/GitHub/Dispatch/Dispatch.iml" />
+      <module fileurl="file:///NikoPC/Documents/GitHub/Dispatch/Dispatch.iml" filepath="/NikoPC/Documents/GitHub/Dispatch/Dispatch.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
-      <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
-      <module fileurl="file://$PROJECT_DIR$/../../../../NikoPC/Documents/GitHub/Dispatch/app/app.iml" filepath="$PROJECT_DIR$/../../../../NikoPC/Documents/GitHub/Dispatch/app/app.iml" />
+      <module fileurl="file:///NikoPC/Documents/GitHub/Dispatch/app/app.iml" filepath="/NikoPC/Documents/GitHub/Dispatch/app/app.iml" />
     </modules>
   </component>
 </project>
