+++ app/src/main/java/com/indexyear/jd/dispatch/activities/MainActivity.java
@@ -1,13 +1,18 @@
 package com.indexyear.jd.dispatch.activities;
 
+import android.content.BroadcastReceiver;
+import android.content.ComponentName;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
+import android.content.IntentFilter;
+import android.content.ServiceConnection;
 import android.content.pm.PackageManager;
 import android.location.Address;
 import android.location.Geocoder;
 import android.location.Location;
 import android.os.Bundle;
+import android.os.IBinder;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.support.design.widget.FloatingActionButton;
@@ -59,6 +64,7 @@ import com.indexyear.jd.dispatch.data.crisis.IGetLatLngListener;
 import com.indexyear.jd.dispatch.data.user.UserManager;
 import com.indexyear.jd.dispatch.models.Crisis;
 import com.indexyear.jd.dispatch.models.User;
+import com.indexyear.jd.dispatch.services.LocationUpdaterService;
 
 import java.util.List;
 import java.util.Locale;
@@ -106,6 +112,12 @@ public class MainActivity extends AppCompatActivity
     //FOR CRISIS
     private String crisisAddress;
 
+    //Luke - FOR LOCATION
+    boolean isBound = false;
+
+    LocationUpdaterService mBoundService;
+    boolean mServiceBound = false;
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         // Start Init
@@ -133,6 +145,7 @@ public class MainActivity extends AppCompatActivity
         SupportMapFragment mapFragment = (SupportMapFragment) getSupportFragmentManager().findFragmentById(map);
         mapFragment.getMapAsync(this);
         getLocationPermission();
+
         // Construct a GeoDataClient.
         mGeoDataClient = Places.getGeoDataClient(this, null);
 
@@ -170,6 +183,13 @@ public class MainActivity extends AppCompatActivity
 
         DEFAULT_ZOOM = 13;
         mDefaultLocation = new LatLng(47.6062, 122.3321);
+
+        //starts service?
+        bindLocationService();
+
+        //registers receiver?
+        newMessage messageReceiver = new newMessage();
+        registerReceiver(messageReceiver, new IntentFilter(NEW_MESSAGE));
     }
 
     private void determineIntent() {
@@ -200,6 +220,14 @@ public class MainActivity extends AppCompatActivity
         } else {
             mUser = getIntent().getParcelableExtra("user");
         }
+    }
+
+    @Override
+    protected void onStart() {
+        super.onStart();
+        Intent intent = new Intent(this, LocationUpdaterService.class);
+        startService(intent);
+        bindService(intent, mConnection, Context.BIND_AUTO_CREATE);
     }
 
     @Override
@@ -489,4 +517,41 @@ public class MainActivity extends AppCompatActivity
     public void onConnectionFailed(@NonNull ConnectionResult connectionResult) {
 
     }
+
+    private void bindLocationService() {
+        try {
+            isBound = getApplicationContext().bindService( new Intent(getApplicationContext(), LocationUpdaterService.class), mConnection, BIND_AUTO_CREATE );
+            bindService(new Intent(this, LocationUpdaterService.class), mConnection, BIND_AUTO_CREATE);
+        } catch (SecurityException e) {
+            // TODO: handle exception
+        }
+    }
+    private ServiceConnection mConnection = new ServiceConnection() {
+        public void onServiceConnected(ComponentName className, IBinder service) {
+            mBoundService = ((LocationUpdaterService.MyBinder)service).getService();
+            Log.d(LocationUpdaterService.TAG, "activity bound to service");
+
+        }
+
+        public void onServiceDisconnected(ComponentName className) {
+
+            mBoundService = null;
+            Log.d(LocationUpdaterService.TAG, "activity unbound to service");
+        }
+    };
+
+
+    //Luke fix this
+    public class newMessage extends BroadcastReceiver
+    {
+        @Override
+        public void onReceive(Context context, Intent intent)
+        {
+            String action = intent.getAction();
+            if(action.equalsIgnoreCase(IMService.NEW_MESSAGE)){
+                Bundle extra = intent.getExtras();
+                double newLat = extra.getDouble("lat");
+                double newLng = extra.getDouble("lng");
+            }
+        }
 }
